Remove IMetricsAccessor otel-extension impl
Fixes #6 by removing the OTel-javaagent extension feature.
Fixes #1.

The IMetricsAccessor interface and related tests remain, as they can be reused in the future for access to IInstrumentManager-metrics. The tests related to the scrape-command have been marked as optional (but permission-tests still remain required), as no IMetricsAccessor implementation exists in this version.